[0.2.5.15] Cleanup connman (#116)
* remove connman options and move startup around

* remove network active toggle

* reduce the amount of work around localnonce

* change eviciton to use data activity instead of other metrics

* keep track of connection failure attempts

* get rid of message maker. less steps same result

* make network_service_version easier to manage

* remove unused banned list changed calls

* use a threadgroup in connman, threadgroup takes in its interrupt varriable as argument

* move network specific files into net folder

* remove message thread specific interruptions

* update formatted files
