+++ src/.formatted-files
@@ -1,7 +1,3 @@
-addrdb.cpp
-addrdb.h
-addrman.cpp
-addrman.h
 args.cpp
 args.h
 amount.cpp
@@ -83,15 +79,20 @@ main.h
 memusage.h
 merkleblock.cpp
 merkleblock.h
-messages.cpp
-messages.h
-net.cpp
-net.h
-netaddress.cpp
-netaddress.h
-netbase.cpp
-netbase.h
-netmessagemaker.h
+net/addrdb.cpp
+net/addrdb.h
+net/addrman.cpp
+net/addrman.h
+net/messages.cpp
+net/messages.h
+net/net.cpp
+net/net.h
+net/netaddress.cpp
+net/netaddress.h
+net/netbase.cpp
+net/netbase.h
+net/protocol.cpp
+net/protocol.h
 networks/netman.cpp
 networks/netman.h
 networks/network.h
@@ -111,8 +112,6 @@ processheader.cpp
 processheader.h
 processtx.cpp
 processtx.h
-protocol.cpp
-protocol.h
 pubkey.cpp
 pubkey.h
 random.cpp
