+++ src/eccoind.cpp
@@ -55,7 +55,8 @@ void WaitForShutdown(thread_group *threadGroup)
 //
 bool AppInit(int argc, char *argv[])
 {
-    thread_group threadGroup;
+    shutdown_threads.store(false);
+    thread_group threadGroup(&shutdown_threads);
 
     bool fRet = false;
