--- src/netmessagemaker.h
@@ -1,50 +1,0 @@
-/*
- * This file is part of the Eccoin project
- * Copyright (c) 2009-2010 Satoshi Nakamoto
- * Copyright (c) 2009-2016 The Bitcoin Core developers
- * Copyright (c) 2014-2018 The Eccoin developers
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#ifndef BITCOIN_NETMESSAGEMAKER_H
-#define BITCOIN_NETMESSAGEMAKER_H
-
-#include "net.h"
-#include "serialize.h"
-
-class CNetMsgMaker
-{
-public:
-    CNetMsgMaker(int nVersionIn) : nVersion(nVersionIn) {}
-    template <typename... Args>
-    CSerializedNetMsg Make(int nFlags, std::string sCommand, Args &&... args) const
-    {
-        CSerializedNetMsg msg;
-        msg.command = std::move(sCommand);
-        CVectorWriter{SER_NETWORK, nFlags | nVersion, msg.data, 0, std::forward<Args>(args)...};
-        return msg;
-    }
-
-    template <typename... Args>
-    CSerializedNetMsg Make(std::string sCommand, Args &&... args) const
-    {
-        return Make(0, std::move(sCommand), std::forward<Args>(args)...);
-    }
-
-private:
-    const int nVersion;
-};
-
-#endif // BITCOIN_NETMESSAGEMAKER_H
