+++ src/rpc/rpcrawtransaction.cpp
@@ -28,7 +28,7 @@
 #include "keystore.h"
 #include "main.h"
 #include "merkleblock.h"
-#include "net.h"
+#include "net/net.h"
 #include "policy/policy.h"
 #include "rpcserver.h"
 #include "script/script.h"
