+++ src/rpc/rpcwallet.cpp
@@ -24,8 +24,8 @@
 #include "core_io.h"
 #include "init.h"
 #include "main.h"
-#include "net.h"
-#include "netbase.h"
+#include "net/net.h"
+#include "net/netbase.h"
 #include "rpcserver.h"
 #include "timedata.h"
 #include "util/util.h"
