+++ src/test/DoS_tests.cpp
@@ -8,7 +8,7 @@
 #include "dosman.h"
 #include "keystore.h"
 #include "main.h"
-#include "net.h"
+#include "net/net.h"
 #include "pow.h"
 #include "script/sign.h"
 #include "serialize.h"
