+++ src/test/addrman_tests.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2015-2017 The Bitcoin Unlimited developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
-#include "addrman.h"
+#include "net/addrman.h"
 #include "test/test_bitcoin.h"
 #include <boost/test/unit_test.hpp>
 #include <string>
