+++ src/test/exploit_tests.cpp
@@ -6,8 +6,8 @@
 
 #include "chain/block.h"
 #include "main.h"
-#include "net.h"
-#include "protocol.h"
+#include "net/net.h"
+#include "net/protocol.h"
 #include "random.h"
 #include "requestManager.h"
 #include "serialize.h"
