+++ src/test/net_tests.cpp
@@ -1,10 +1,10 @@
 // Copyright (c) 2012-2016 The Bitcoin Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
-#include "net.h"
-#include "addrman.h"
+#include "net/net.h"
 #include "crypto/hash.h"
-#include "netbase.h"
+#include "net/addrman.h"
+#include "net/netbase.h"
 #include "serialize.h"
 #include "streams.h"
 #include "test/test_bitcoin.h"
