+++ src/test/netbase_tests.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "netbase.h"
+#include "net/netbase.h"
 #include "test/test_bitcoin.h"
 #include "util/utilstrencodings.h"
