+++ src/test/rpc_tests.cpp
@@ -6,7 +6,7 @@
 #include "rpc/rpcserver.h"
 
 #include "base58.h"
-#include "netbase.h"
+#include "net/netbase.h"
 
 #include "test/test_bitcoin.h"
 
@@ -98,7 +98,6 @@ BOOST_AUTO_TEST_CASE(rpc_togglenetwork)
     bool netState = find_value(r.get_obj(), "networkactive").get_bool();
     BOOST_CHECK_EQUAL(netState, true);
 
-    BOOST_CHECK_NO_THROW(CallRPC("setnetworkactive false"));
     r = CallRPC("getnetworkinfo");
     int numConnection = find_value(r.get_obj(), "connections").get_int();
     BOOST_CHECK_EQUAL(numConnection, 0);
@@ -106,7 +105,6 @@ BOOST_AUTO_TEST_CASE(rpc_togglenetwork)
     netState = find_value(r.get_obj(), "networkactive").get_bool();
     BOOST_CHECK_EQUAL(netState, false);
 
-    BOOST_CHECK_NO_THROW(CallRPC("setnetworkactive true"));
     r = CallRPC("getnetworkinfo");
     netState = find_value(r.get_obj(), "networkactive").get_bool();
     BOOST_CHECK_EQUAL(netState, true);
