+++ src/test/test_bitcoin.cpp
@@ -16,7 +16,7 @@
 #include "init.h"
 #include "key.h"
 #include "main.h"
-#include "messages.h"
+#include "net/messages.h"
 #include "processblock.h"
 #include "pubkey.h"
 #include "random.h"
