+++ src/test/test_bitcoin_fuzzy.cpp
@@ -6,13 +6,13 @@
 #include "config/bitcoin-config.h"
 #endif
 
-#include "addrman.h"
 #include "chain/chain.h"
 #include "coins.h"
 #include "compressor.h"
 #include "consensus/merkle.h"
-#include "net.h"
-#include "protocol.h"
+#include "net/addrman.h"
+#include "net/net.h"
+#include "net/protocol.h"
 #include "pubkey.h"
 #include "script/script.h"
 #include "streams.h"
