+++ src/wallet/walletdb.cpp
@@ -23,8 +23,8 @@
 #include "args.h"
 #include "base58.h"
 #include "consensus/validation.h"
+#include "net/protocol.h"
 #include "processtx.h" // For CheckTransaction
-#include "protocol.h"
 #include "serialize.h"
 #include "sync.h"
 #include "util/util.h"
